requirements.txt
django-bootstrap3
django-allauth